Major cleanup to game interface.  Added multiple helper methods for optimizing dispaly.
1.) This includes a 'FormatTextWidth' method that breaks apart a "sentance" or "paragraph" String into it's component words (splitting on spaces) and taking the global width integer (defines the number of textual character columns) so that the text can be fed in as a single String, broken apart into logical sentances that fit within the maximum width, and are collected as an array.  They are than reassembled into a String with new-line breaks so that words are preserved.  No support for hyphenation is being considered.

2.) Created the master game loop that all actions will happen within.  It will continue as a while loop until the "endOfGame" value is set.  At which point the score will be given.

3.) Added additional text descriptions to make the experience richer.  Also formatting it all through the new 'FormatTextWidth' method for consistency, ensuring a leading space is used for proper padding in the console window.

4.) Also added a colorizing method so that room titles and other important messages can be displayed in the console color of choice, without the extra ConsoleColor sets and resets muddying up the code.  Easy now to pass in the desired ConsoleColor enumeration and the text to colorize!

5.)  All of that is in the main code.  I have made additional improvements in the Player and RoomLocation classes to support the changes and new methods in the MisersHouse Main method.

6.)  Added a standardized "showRoomInfo(...)" method that clears the screen so that rooms are displayed cleanly with each room movement the player makes.  The room is described, and any objects (future code) will be shown here.

I shall need to convert to uppercase on the next commit it seems... I am breaking C# standards by camel casing method names, whereas camelcase happens to be the standard for the Java!  Takes a while to transition ones mind completely!

7.) Added a method for building the compass directions available to each player.  This is a pretty cool little method, I think!  Quick and easy to grab the array out of the object, and by creating a quick array of matching, directions (N,S,E,W) it allows me to easily display only the directions with a valid room value associated with it.  Zero values mean that the direction is not valid, and I don't want to display it.  Works well!

This game is coming together a lot faster than I expected, since I have little prior experience in C# or the .NET framework... just in Java.  Close languages, actually, which makes it fun, but definite diffrences between them.

I find it exciting that often when debugging, I find my issues quickly (granted, I am writing it!), but also fun that I have had many epiphany moments so far, in which I try something, and it works the first time!  I try out a new statement (even a complex one), and I can often make it work on the first try!

The case statement change (from the cluttered series of else / 'if else' structure) was tricky, because I was performing a string comparison and converting using a rather wordy "StringComparison.InvariantCultureIgnoreCase()" comparison, so I was able to change it, relatively easily, to a case statement that handled the comparison while also eliminating complexity.

I've learned a ton about the C# language working on this little pet project, and I hope to do some amazing things with regards to the language processing as well!