Converting some of my colorized console statements to use the method for handling color changes.
Also trying to figure out why it is not identifying me as ME.  ;-)  I added in my e-mail address so that it shoudl have identified me.  But still I am committing across as a doppleganger me thinks.

This commit, I changed my username from 'Michael Monson' to 'michaelmonson'

We'll see if that makes a difference. :-)