General game changes to expand rooms and fix directional logic.    1. Increased the accessible rooms from 6 to 48.    2. Added 'hug' as a valid verb, though still missing logic.    3. Added "LOOK" logic (now calls location description method; refreshes view.    4. Added directions for South(S), East(E), West(W) and associated logic.  Also re-displaying available directions for player.    5. Increased "I Don't Know" variations for the game oracle.    6. Also added a random set of six responses for "wrong way / invalid direction" logic.    7. Made the 'buildCompassDirections(...) method public, so it can be re-used from other places, such as the languageParser.    8. Added a new utilities class for assorted "helper" methods.  However, I'm not able to use the first method quite properly yet... which is a wrapper object for simplifying ENUM subset searches.  I'd like to add in an overall check for ALL direections, rather than chainging them together with a series of OR || logical operations.
Many more changes to come.  The next big change will be to add in an "object" object.  ;-)  A structure that will enable defining objects that can be interacted with, as well as their state (visible, hidden, part of your inventory, damaged, lost, etc.)

Writing this is a ton of fun!  Of course, some of the experiments being made by teams of programmers to use AI and machine learning will likely make these sorts of adventures obsolete.  But it is still worth the time to programmatically execute a game like this.  It is a good exercise in logic!  And great for learning a programming langauge in a fun, entertaining way!