Added a new Inventory class for tracking what a player carries.    - A whole slew of changes with this commit!  It's getting more and more intelligent all of the time, though it is taking a while because all of the infrastructure needs to be built up first.    - I'm chopping at the bit because there is so much language functionality I want to build into the paser, to make the games understanding of language rich and flexible, as to make this a rewarding (not frustrating) experience for the player!  But I need to hold off on that until the basic founcations are in place.    - But I dream at night about what I would like to see the game do, and I think through logic and enjoy epiphanies as I shower in the morning. :-)    - Here are the features and changes with this commit:    1.)  Added the new "Inventory.cs" class
   2.)  Updated the necessary GameItem constructors and property lists to support the new property.

   3.)  Added a new object state (CLOSED)

   4.)  Added the base "name" to the object properties that exist for easier searching... though now that I have learned about "CONTAINS" it seems reduntant to have an extra object propety when the "base" part of the name can simply be searched within the ItemId constant.  In fact, it may even be more flexible, since "WATERBOTTLE" can handle a player search of both "water," "bottle," or "water bottle" (with some tweaking)  - Thanks to Contains()

   5.) Adding a new "NameDetailed" property to the GameItem objects, so that I can supply a more descriptive name for the item that was found (e.g. "brass door key" instead of key, and "juicy mango" instead .
of mango) and for displaying in the inventory and other lists and contexts.

   6.) LANGUAGE PARSER - tons of changes!

Added the concept of a "LastCommand" for handling context of the previous player command.  In this case, the game oracle may be asking for clarification "get what?' or "talk to whom?" or "get which bottle, the blue bottle or the green one?"

Expanded the list of Verbs significantly.  Also added an enumeration of Prepositions, Articles & Determiners, and basic Possessives.  I'm re-learning some rules of English all over! :-)

Added a REGEX for filtering out pronunciation marks, such as explanation marks and question marks.

Language Logic added to more accurately identify the noun.  Still need work for handling adjectives.  I tell you, it is like internally diagraming sentances!

Added additional commands:  LOOK, WHO [are ____ ], directional improvements, GET (e.g. TaKE PICK UP, ACQUIRE), and more.

Made certain properties in the MisersHouseMain into static values for access throughout the game.  Globals... yeah!