Added an "Inventory Intializer" method that I now call from the program main.    - Much better!  The way it shoudl have been.    - Inventory now correctly shows all items.    - Because a single "Game Items" collection exists for ALL items, it's easy to display matches and move items between each necessary status.    - One basket for all the game items is preferable!    - Also consolidated several methods    - Eliminated excess / outdated comments.
Keeps moving me in the direction I need to go as I build up the business logic and framework of the Application.  Once I have a few more command in, I shoudl be able to start coding the rest of the objects in, as well as which room they apper in, and how the player interacts with each of them!