Update MisersHouseMain.cs to use a verbatim string literal
This is an alternative approach that will make it MUCH easier to use ASCII Art in my projects, since it doesn't required escaped characters.

In a verbatim string literal (indicated by the @ symbol in C#), backslashes do not need to be escaped.  This approach is perfect for multi-line ASCII art or text blocks. It’s simple and clean, avoiding the hassle of escaping characters. 

In other languages, such as JavaScript, and more recently with Java, there is the "triple quote" or "Comment Block notation. 

At any rate, it can be tried out,  and see where that takes me.  :-)